app/api/views.py
"""API endpoints for system operations"""
from rest_framework import status
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework.request import Request

from .serializers import company, members, offers
from .handlers import handle_webhook
from core.utils.error_handler import ErrorHandler


@api_view(['POST'])
def webhook_endpoint(request: Request) -> Response:
    """Handle incoming webhooks"""
    # Get webhook type
    webhook_type = request.data.get('metadata', {}).get('event_type')
    if not webhook_type:
        error = ErrorHandler.handle_system_error(
            code="MISSING_TYPE",
            service="webhook",
            action="validate",
            message="Missing webhook type"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # Process webhook
    result = handle_webhook(webhook_type, request.data)
    if "error" in result:
        return Response(result, status=status.HTTP_400_BAD_REQUEST)
    return Response(result, status=status.HTTP_200_OK)


@api_view(['GET'])
def list_companies(request: Request) -> Response:
    """List companies with optional filters"""
    # Validate filters
    serializer = company.CompanySerializer(data=request.query_params)
    if not serializer.is_valid():
        error = ErrorHandler.handle_system_error(
            code="INVALID_FILTERS",
            service="company",
            action="validate_filters",
            message="Invalid filter parameters"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement company listing logic
    companies = []  # Replace with actual company fetching
    return Response(companies, status=status.HTTP_200_OK)


@api_view(['GET'])
def get_company(request: Request, company_id: str) -> Response:
    """Get company details"""
    if not company_id:
        error = ErrorHandler.handle_system_error(
            code="MISSING_ID",
            service="company",
            action="validate",
            message="Company ID is required"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement company retrieval logic
    company_data = {}  # Replace with actual company fetching
    return Response(company_data, status=status.HTTP_200_OK)


@api_view(['GET'])
def list_members(request: Request) -> Response:
    """List members with optional filters"""
    # Validate filters
    serializer = members.MemberSerializer(data=request.query_params)
    if not serializer.is_valid():
        error = ErrorHandler.handle_system_error(
            code="INVALID_FILTERS",
            service="member",
            action="validate_filters",
            message="Invalid filter parameters"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement member listing logic
    members_list = []  # Replace with actual member fetching
    return Response(members_list, status=status.HTTP_200_OK)


@api_view(['GET'])
def get_member(request: Request, member_id: str) -> Response:
    """Get member details"""
    if not member_id:
        error = ErrorHandler.handle_system_error(
            code="MISSING_ID",
            service="member",
            action="validate",
            message="Member ID is required"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement member retrieval logic
    member_data = {}  # Replace with actual member fetching
    return Response(member_data, status=status.HTTP_200_OK)


@api_view(['GET'])
def list_offers(request: Request) -> Response:
    """List offers with optional filters"""
    # Validate filters
    serializer = offers.OfferSerializer(data=request.query_params)
    if not serializer.is_valid():
        error = ErrorHandler.handle_system_error(
            code="INVALID_FILTERS",
            service="offer",
            action="validate_filters",
            message="Invalid filter parameters"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement offer listing logic
    offers_list = []  # Replace with actual offer fetching
    return Response(offers_list, status=status.HTTP_200_OK)


@api_view(['GET'])
def get_offer(request: Request, offer_id: str) -> Response:
    """Get offer details"""
    if not offer_id:
        error = ErrorHandler.handle_system_error(
            code="MISSING_ID",
            service="offer",
            action="validate",
            message="Offer ID is required"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement offer retrieval logic
    offer_data = {}  # Replace with actual offer fetching
    return Response(offer_data, status=status.HTTP_200_OK)


@api_view(['POST'])
def accept_offer(request: Request, offer_id: str) -> Response:
    """Accept an offer"""
    if not offer_id:
        error = ErrorHandler.handle_system_error(
            code="MISSING_ID",
            service="offer",
            action="validate",
            message="Offer ID is required"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement offer acceptance logic
    return Response(
        {"message": "Offer accepted successfully"},
        status=status.HTTP_200_OK
    )


@api_view(['POST'])
def reject_offer(request: Request, offer_id: str) -> Response:
    """Reject an offer"""
    if not offer_id:
        error = ErrorHandler.handle_system_error(
            code="MISSING_ID",
            service="offer",
            action="validate",
            message="Offer ID is required"
        )
        return Response(error, status=status.HTTP_400_BAD_REQUEST)

    # TODO: Implement offer rejection logic
    return Response(
        {"message": "Offer rejected successfully"},
        status=status.HTTP_200_OK
    )
