requirements/base.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
# Updated to fix CVE-2024-22513 (information disclosure vulnerability)
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
django-environ==0.11.2

# Database and Caching
hiredis==3.1.0
redis==5.2.1
async-timeout==5.0.1

# Utils
python-decouple==3.8
python-json-logger==3.2.1
python-logstash==0.4.8
requests==2.32.3
# urllib3 constraint matches requirements from multiple dependencies:
# - requests requires >=1.21.1<3
# - types-requests requires >=2
# - watchtower->boto3->botocore requires >=1.25.4<3!=2.2.0
urllib3>=2.0.0,<3.0.0
watchtower==3.3.1
PyJWT==2.10.1

# Dependencies
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
idna==3.10
importlib_metadata==8.5.0
Markdown==3.7
sqlparse==0.5.3
typing_extensions==4.12.2
zipp==3.21.0
